Delete functionality added
list.ejs file:  Added a form arounf checkbox and submit functionality to
delete items. The form data is sent to "/delete" and checkbox has the
value of item _id to be accessed later

app.js file: Here we access checkbox item _id in post request for
"/delete" and we delete the particular entry from database and redirect
to homepage for the user to see changes.

style.css: A small change made in form class added specific item to be
targeted.